fcmp++ TreeSync: fix impl bugs, tests passing
- When trimming tree, always re-grow the last chunk with remaining
elems left in the chunk, since we aren't guaranteed to have all
the elems from the chunk that need to be explicilty trimmed.
- Make sure to update existing last hashes every block, so that
upon trimming, it's using the latest state for new last hashes
and trimming as expected accordingly.
- Update ref'd last hashes using the tree reduction always upon
popping block, not only when updating registered output paths.
- Once a leaf is removed from the tree, can remove refs to all
chunk elems in the leaf's path, including elems that might still
exist in the tree. Don't need to keep track of them on behalf of
the registered output anymore.
- Lots of cosmetic changes.