fcmp++: fix tree cache reorg handling after syncing n blocks
Since we sometimes sync multiple blocks at a time, and syncing
is optimized by constructing a single tree extension, in order to
trim to a prior tree edge correctly, we need to make sure we're
using the correct version of the prior tree edge. We can either do
more hashes when forward syncing to calculate each block's tree
edge, or we can calculate the expected tree edge on reorg. This
commit opts for the latter to optimize forward sync.